Unify optimize C++ functions.
